+++ projects/np-ui-lib/src/lib/np-timeline/np-timeline.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NpTimelineComponent } from './np-timeline.component';
+
+@NgModule({
+  declarations: [NpTimelineComponent],
+  imports: [
+    CommonModule
+  ],
+  exports: [NpTimelineComponent]
+})
+export class NpTimelineModule { }