+++ src/app/np-timeline-demo/np-timeline-demo-routing.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { NpTimelineDemoComponent } from './np-timeline-demo.component';
+
+const routes: Routes = [{
+  path: '',
+  component: NpTimelineDemoComponent
+}];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class NpTimelineDemoRoutingModule { }